chore: improve Avatar, Alert, AlertDialog, Avatar exports
